requirements.txt
asgiref==3.3.1
attrs==20.3.0
bcrypt==3.2.0
cached-property==1.5.2
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
cryptography==3.2.1
distro==1.5.0
Django==3.1.4
django-phonenumber-field==5.0.0
djangorestframework==3.12.2
docker==4.4.0
docker-compose==1.27.4
dockerpty==0.4.1
docopt==0.6.2
future==0.18.2
idna==2.10
jsonschema==3.2.0
paramiko==2.7.2
phonenumbers==8.12.13
Pillow==8.0.1
psycopg2==2.8.6
pycparser==2.20
PyNaCl==1.4.0
pyrsistent==0.17.3
python-dotenv==0.15.0
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
texttable==1.6.3
urllib3==1.26.2
websocket-client==0.57.0
